Simple HTTP Request with OkHttp - Android Studio
we will use the OkHttp library to make a simple asynchronous HTTP request, download a JSON from a URL and display it in a TextView. For this we will use OkHttp's enqueue method which executes the network call in a background thread. To update the TextView we will then access our main thread with the runOnUiThread method. JSON courtesy of Reqres A hosted REST-API ready to respond to your AJAX requests.